A dialogue box appears when deleting, for confirmation

+++ SecurePass/SecurePass/Pages/DetailsPage.xaml.cs
@@ -15,6 +15,7 @@ namespace SecurePass.Pages
     {
         public DetailsPage()
         {
+            
             InitializeComponent();
 
             var nameEntry = new Entry();
@@ -23,23 +24,41 @@ namespace SecurePass.Pages
             var passEntry = new Entry();
             passEntry.SetBinding(Entry.TextProperty, "Password");
 
+           
+
             var doneSwitch = new Switch();
             doneSwitch.SetBinding(Switch.IsToggledProperty, "Done");
 
             var saveButton = new Button { Text = "Save" };
             saveButton.Clicked += async (sender, e) =>
             {
-                var application = (User)BindingContext;
-                await App.Database.SaveApplicationAsync(application);
-                await Navigation.PopAsync();
+                if (nameEntry.Text.Length == 0)
+                {
+                    await DisplayAlert("Oops", "Please enter an account name", "OK");
+                }
+                else
+                {
+                    var application = (User)BindingContext;
+                    await App.Database.SaveApplicationAsync(application);
+                    await Navigation.PopAsync();
+                }
+                
             };
 
             var deleteButton = new Button { Text = "Delete" };
             deleteButton.Clicked += async (sender, e) =>
-            {
-                var application = (User)BindingContext;
-                await App.Database.DeleteItemAsync(application);
-                await Navigation.PopAsync();
+            {                
+                {
+                    var result =
+                      await DisplayAlert("Confirmation", "Are you sure? This cannot be undone",
+                      "OK", "Cancel");
+                    if (result == true)
+                    {
+                        var application = (User)BindingContext;
+                        await App.Database.DeleteItemAsync(application);
+                        await Navigation.PopAsync();
+                    }
+                }
             };
 
             var cancelButton = new Button { Text = "Cancel" };
@@ -48,7 +67,8 @@ namespace SecurePass.Pages
                 await Navigation.PopAsync();
             };
 
-            
+
+
 
             Content = new StackLayout
             {
@@ -79,20 +99,28 @@ namespace SecurePass.Pages
                 return;
             }
         }
+        string nameEntry { get; set; }
+        string passEntry { get; set; }
+
+
 
 
         async void OnSaveClicked(object sender, EventArgs e)
         {
+            validate(nameEntry, passEntry);
             var application = (User)BindingContext;
             await App.Database.SaveApplicationAsync(application);
             await Navigation.PopAsync();
+           
         }
 
-        async void OnDeleteClicked(object sender, EventArgs e)
+        private async void OnDeleteClicked(object sender, EventArgs e)
         {
-            var application = (User)BindingContext;
-            await App.Database.DeleteItemAsync(application);
-            await Navigation.PopAsync();
+            
+                    var application = (User)BindingContext;
+                    await App.Database.DeleteItemAsync(application);
+                    await Navigation.PopAsync();               
+            
         }
 
         async void OnCancelClicked(object sender, EventArgs e)
@@ -100,5 +128,18 @@ namespace SecurePass.Pages
             await Navigation.PopAsync();
         }
 
+        async void validate(string accountName, string password)
+        {
+            if (!string.IsNullOrWhiteSpace(accountName) && !string.IsNullOrWhiteSpace(password))
+            {
+                User user = new User
+                {
+                    AccountName = accountName
+
+                };
+                await DisplayAlert("Oops", "Please enter an account name", "OK");
+
+            }
+        }
     }
 }
